Tweak the output logs a bit
